refactor(extension): rename extensions to extension
